more updation with numpy tsk

+++ graphs/models.py
@@ -20,43 +20,43 @@ class book1(models.Model):
 
 
 class book2(models.Model):
-    value1 = models.IntegerField()
-    value2 = models.IntegerField()
-    value3 = models.IntegerField()
-    value4 = models.IntegerField()
-    value5 = models.IntegerField()
-    value6 = models.IntegerField()
-    value7 = models.IntegerField()
-    value8 = models.IntegerField()
-    value9 = models.IntegerField()
-    value10 = models.IntegerField()
+    value1 = models.DecimalField(max_digits=5, decimal_places=5)
+    value2 = models.DecimalField(max_digits=5, decimal_places=5)
+    value3 = models.DecimalField(max_digits=5, decimal_places=5)
+    value4 = models.DecimalField(max_digits=5, decimal_places=5)
+    value5 = models.DecimalField(max_digits=5, decimal_places=5)
+    value6 = models.DecimalField(max_digits=5, decimal_places=5)
+    value7 = models.DecimalField(max_digits=5, decimal_places=5)
+    value8 =models.DecimalField(max_digits=5, decimal_places=5)
+    value9 = models.DecimalField(max_digits=5, decimal_places=5)
+    value10 = models.DecimalField(max_digits=5, decimal_places=5)
     
 
 class book3(models.Model):
-    value1 = models.IntegerField()
-    value2 = models.IntegerField()
-    value3 = models.IntegerField()
-    value4 = models.IntegerField()
-    value5 = models.IntegerField()
-    value6 = models.IntegerField()
-    value7 = models.IntegerField()
-    value8 = models.IntegerField()
-    value9 = models.IntegerField()
-    value10 = models.IntegerField()
+    value1 = models.DecimalField(max_digits=5, decimal_places=5)
+    value2 = models.DecimalField(max_digits=5, decimal_places=5)
+    value3 = models.DecimalField(max_digits=5, decimal_places=5)
+    value4 = models.DecimalField(max_digits=5, decimal_places=5)
+    value5 = models.DecimalField(max_digits=5, decimal_places=5)
+    value6 = models.DecimalField(max_digits=5, decimal_places=5)
+    value7 = models.DecimalField(max_digits=5, decimal_places=5)
+    value8 =models.DecimalField(max_digits=5, decimal_places=5)
+    value9 = models.DecimalField(max_digits=5, decimal_places=5)
+    value10 = models.DecimalField(max_digits=5, decimal_places=5)
     
 
 
 class book4(models.Model):
-    value1 = models.IntegerField()
-    value2 = models.IntegerField()
-    value3 = models.IntegerField()
-    value4 = models.IntegerField()
-    value5 = models.IntegerField()
-    value6 = models.IntegerField()
-    value7 = models.IntegerField()
-    value8 = models.IntegerField()
-    value9 = models.IntegerField()
-    value10 = models.IntegerField()
+    value1 = models.DecimalField(max_digits=5, decimal_places=5)
+    value2 = models.DecimalField(max_digits=5, decimal_places=5)
+    value3 = models.DecimalField(max_digits=5, decimal_places=5)
+    value4 = models.DecimalField(max_digits=5, decimal_places=5)
+    value5 = models.DecimalField(max_digits=5, decimal_places=5)
+    value6 = models.DecimalField(max_digits=5, decimal_places=5)
+    value7 = models.DecimalField(max_digits=5, decimal_places=5)
+    value8 =models.DecimalField(max_digits=5, decimal_places=5)
+    value9 = models.DecimalField(max_digits=5, decimal_places=5)
+    value10 = models.DecimalField(max_digits=5, decimal_places=5)
